Refactor evaluator models and remove profiler test.
Move evaluator-related models from `nat.eval.evaluator` to `nat.data_models` to improve module organization. Remove outdated `test_profiler.py` and update documentation to include details on the new `prediction_trie` functionality.
